[Example] Move golden to test file, simplify example smoke test
- Golden (ref_fwd_varlen/ref_bwd_varlen) and check_precision moved to test file
- Example file: smoke test simplified to single case + non-trivial output check
- Test file: precision tests only (L0/L1/L2/Boundary), removed do_bench/msprof
